weather/requirements.txt
numpy
python_twitter
pypng
scipy
PIL
python-dateutil
requests
